Add timeout to HTTP requests
By default, there is no timeout. So when there is nothing to connect to, there could be a long waiting period until the connection itself was closed.

Added a default of 3 seconds. This way there is a much faster response. Since printers will be mostly on local networks this is more than enough time.